requirements.txt
flask==3.0.3
anthropic==0.34.0
reportlab==4.2.2
google-auth==2.34.0
google-api-python-client==2.143.0
gunicorn==23.0.0
